d3d11?

+++ src/graphic/Fast3D/gfx_direct3d11.cpp
@@ -1230,7 +1230,7 @@ gfx_d3d11_get_pixel_depth(int fb_id, const std::set<std::pair<float, float>>& co
 } // namespace
 
 ImTextureID gfx_d3d11_get_texture_by_id(unsigned long long id) {
-    return d3d.textures[id].resource_view.Get();
+    return (unsigned long long)d3d.textures[id].resource_view.Get();
 }
 
 void gfx_d3d11_enable_srgb_mode() {
